fix(test): drop unused oss-cluster flag; extract real cluster class to own file
testGraphMemoryUsage (test_memory_usage.py:44) was constructed with
env='oss-cluster' but never used any cluster feature — no shardId, no
getConnection(shardId=...), no master_N. Under RLTest the flag spawned a
local multi-process cluster and the test happened to talk to the master;
behavior is identical against a single node. Image-based CI doesn't
implement cluster spawn (Mode 3's _spawn_falkordb is single-container),
so the cluster guard rightly fires and the whole file fails. Drop the
flag — the test still tests what it always tested.

testUDFCluster (test_udf.py:1575+) genuinely needs cluster topology:
it opens getConnection(shardId=1/2/3) separately and tests that
GRAPH.UDF.LOAD/DELETE/FLUSH propagate across master shards. Single-
node would make those propagation assertions vacuous. Extract to
tests/flow/test_udf_cluster.py so the rest of test_udf.py (testUDF +
test_udf_javascript, neither cluster) can route to services and pass.
The new file routes to spawn via CLUSTER_RE and still fails with the
"cluster not implemented" RuntimeError; that failure is now isolated
to one file rather than dragging two unrelated files down with it.

Implementing real cluster orchestration in spawn-mode (spawn N
containers, CLUSTER MEET, slot assignment, getConnection(shardId)
glue) is a follow-up.
