+++ flow_tests_done.txt
@@ -96,6 +96,7 @@ tests/flow/test_temporal
 tests/flow/test_timeout.py
 tests/flow/test_traversal_construction.py
 tests/flow/test_udf.py
+tests/flow/test_udf_cluster.py
 tests/flow/test_undo_log.py
 tests/flow/test_union
 tests/flow/test_unwind_clause
